addresses code organization & comments
